requirements.txt
alembic==1.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.11.0
backoff==2.2.1
bleach==6.2.0
blurhash==1.1.5
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
coloredlogs==14.0
confluent-kafka==2.12.0
dagit==1.11.14
dagster==1.11.14
dagster-graphql==1.11.14
dagster-pipes==1.11.14
dagster-webserver==1.11.14
dagster_shared==1.11.14
decorator==5.2.1
docstring_parser==0.17.0
dotenv==0.9.9
filelock==3.20.0
fsspec==2025.9.0
gql==3.5.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
grpcio==1.75.1
grpcio-health-checking==1.75.1
h11==0.16.0
httptools==0.7.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
kafka==1.3.5
kafka-python==2.2.15
kaggle==1.7.4.5
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
Mastodon.py==2.1.4
mdurl==0.1.2
multidict==6.7.0
numpy==1.26.4
packaging==25.0
pandas==1.5.3
pathlib_abc==0.5.1
propcache==0.4.1
protobuf==6.32.1
psycopg2-binary==2.9.11
py4j==0.10.9.9
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
pyspark==4.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic==0.4.27
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
scikit-learn==1.2.2
scipy==1.11.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
structlog==25.4.0
tabulate==0.9.0
text-unidecode==1.3
threadpoolctl==3.6.0
tomli==2.3.0
tomlkit==0.13.3
toposort==1.10
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
universal_pathlib==0.3.3
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
yarl==1.22.0
